# src/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.3.6"
  title: BetProtocol API
# during dev, should point to your local machine
host: localhost:8080
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /users/register:
    x-swagger-router-controller: user
    post:
      tags:
      - user
      summary: Register User Data
      operationId: registUser
      description: Search for matching accounts in the system.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          description: User object to submit to the network
          required: true
          schema:
            $ref: '#/definitions/UserRegisterRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/login:
    x-swagger-router-controller: user
    post:
      tags:
      - user
      summary: Login User
      operationId: loginUser
      description: Login User Data
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          description: User object to submit to the network
          required: true
          schema:
            $ref: '#/definitions/UserLoginRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/bets: 
    x-swagger-router-controller: user
    post:
      tags:
      - user
      - bets
      summary: Get all Bets for the User
      operationId: getBets
      description: Get all Bets for the User
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          description: User Information
          required: true
          schema:
            $ref: '#/definitions/UserBetsRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/summary:
    x-swagger-router-controller: user
    post:
      tags:
      - user
      summary: Get Summary Data for User
      operationId: userSummary
      description: Get Summary Data for User
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          description: User Information
          required: true
          schema:
            $ref: '#/definitions/UserSummaryRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/updateWallet: 
    x-swagger-router-controller: user
    post:
      tags:
      - user
      summary: Update Wallet Amount for the User
      operationId: updateWallet
      description: Update Wallet Amount for the User
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          description: User Information
          required: true
          schema:
            $ref: '#/definitions/UpdateUserWalletRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/finalizeWithdraw:
    x-swagger-router-controller: user
    post:
      tags:
      - user
      summary: Request for User Finalize Withdraw
      operationId: finalizeWithdraw
      description: Request Finalize Withdraw
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          description: User Information
          required: true
          schema:
            $ref: '#/definitions/UserFinalizeWithdrawRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /app/summary:
    x-swagger-router-controller: app
    post:
      tags:
      - app
      summary: Get Summary Data for App
      operationId: summary
      description: Get Summary Data for App
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: app
          description: App Information
          required: true
          schema:
            $ref: '#/definitions/AppSummaryRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  UserRegisterRequest: 
    type: object
    required:
      - username
      - name             
      - email
      - address 
      - app            
    properties:
      username:
        type : string
        format: username
        example: "Jamy_234"
        minLength: 1
        description: |
          Unique identifier of the user, besides the ID
      name:
        type : string
        format: name
        example: "James Rockforth"
        minLength: 1
        description: |
          User Real Name
      email:
        type : string
        format: email
        example: "jamy@gmail.com"
        minLength: 1
        description: |
          User Email
      app:
        type : string
        example: "9a7yfg8sd7fygs87"
        minLength: 1
        description: |
          App Id
      address:
        type : string
        example: 0x98g2giqehr9vn8342
        minLength: 1
        description: |
          User Address
...
  UserLoginRequest: 
    type: object
    required:
      - username
      - password 
      - app
    properties:
      username:
        type : string
        format: username
        example: "Jamy_234"
        minLength: 1
        description: |
          Unique identifier of the user, besides the ID
      password:
        type : string
        format: password
        example: "$ufhfwhe)#9grwg"
        minLength: 1
        description: |
          User Password Encrypted
      app:
        type : string
        example: 53vi230283vv234fg2v37243
        description: |
          Unique identifier of the App Id
  UserSummaryRequest:
    type: object
    required:
      - type
      - user
    properties:
      type:
        type : string
        format: name
        example: "Revenue"
        minLength: 1
        maxLength: 20
        description: |
          Type of Summary Data
      user :
        type : string
        format: name
        example: "0xhi834ty2034thg"
        description: |
          User ID
...
  UserFinalizeWithdrawRequest : 
    type: object
    required:
      - app
      - user
      - address
      - signature
      - tokenAmount
      - transactionHash
      - withdraw_id
    properties:
      tokenAmount:
        type : number
        example: 203
        description: |
          New User Balance for User
      app :
        type : string
        format : name
        example : "5cuedgq7e6t334t34" 
        description: |
          App ID
      user :
        type : string
        format: name
        example : "5cuedgq7e6t334t34" 
        description: |
          User ID
      address :
        type : string
        format: name
        example: "0xe8rghwe8rghwe9r87ynw987eyncwre"
        description: |
          User Address
      signature :
        type : object
        description: |
          User Signature 
      transactionHash :
        type : string
        format: name
        example: "0xe8rghwe8rghwe9r87ynw987eyncwre"
        description: |
          Transaction Hash
      withdraw_id :
        type : string
        format: name
        example: "65f76st8vdsb3181"
        description: |
            Withdraw Id
  UserBetsRequest : 
    type: object
    required:
      - user
      - size
    properties:
      size:
        type : number
        example: 15
        description: |
          Amount of Bets Wanted
      user :
        type : string
        format: name
        example: "0xhi834ty2034thg"
        description: |
          User ID
  UpdateUserWalletRequest : 
    type: object
    required:
      - user
      - amount
      - transactionHash
      - app
    properties:
      amount:
        type : number
        example: 203
        description: |
          Amount added or taken from Wallet
      user :
        type : string
        format: name
        example: "0xhi834ty2034thg"
        description: |
          User ID
      app :
        type : string
        format: name
        example: "0xhi834ty2034thg"
        description: |
          App ID
      transactionHash :
        type : string
        format: name
        example: "0xe8rghwe8rghwe9r87ynw987eyncwre"
        description: |
          Transaction Hash
  AppSummaryRequest:
    type: object
    required:
      - type
      - app
    properties:
      type:
        type : string
        format: name
        example: "Bet.pt"
        minLength: 1
        maxLength: 20
        description: |
          App Name
      app :
        type : string
        format: name
        example: "0xhi834ty2034thg"
        description: |
          App ID
      periodicity :
        type : string
        format: name
        example: "Weekly"
        description: |
         Periodicity of Data
...
  GeneralResponse : 
    type: object
    required:
      - data
      - meta            
    properties:
      data:
        type : object
        format: data
        description: |
          Unique identifier of the user, besides the ID
      meta:
        type : object
        format: meta
        description: |
          User Real Name
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string